Fix Kudu scanner KeepAlive with large Kudu batches
On a cluster with 32 cores per node, I kept hitting "Scanner not found"
failures when querying COUNT(*) on a TPCH 15TB lineitem table. The issue
was that Impala was not issuing scanner KeepAlive RPCs to Kudu often
enough.

The previous implementation of KeepAlive issued a KeepAlive whenever a scanner
thread blocked for 15 seconds trying to add a row batch to the rowbatch queue.
However, in the case that there are many scanner threads (in this case 48
because of the high core count) and the Kudu-side scan batches are large (in
this case 8-10M per batch, because of the empty COUNT(*) projection), we end up
with a scenario where all of the scanner threads are round-robin feeding Impala
row batches to the main thread, but only rarely need to go back to Kudu to grab
a new Kudu batch. Because every scanner was making progress from the Impala
perspective, we didn't send a KeepAlive, and Kudu timed out the scanners.

The fix is to keep track of the last time at which we've either sent a keepalive
or made a successful scan RPC to Kudu, and after every batch, check the timer.
The cost is an extra clock syscall per rowbatch, but this is likely in the noise
since we already read the clock many times per batch to generate timer counters.

This bug is hard to repro in a unit test, but I verified that I'm now able to
reliably run the COUNT query on the large dataset. This used to fail reliably.

Change-Id: I4bdceed5e73a260c0caacace34efb17351d54633
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/10038
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins