Fix uninitialized class member in KuduScanner
The rows_scanned_current_block_ member was uninitialized.  I saw this
cause a crash in one cluster run, though it's hard to deterministically
reproduce. Looking at the core file, I found that
rows_scanned_current_block_ was -1319175936, which could likely cause
incorrect results.

This likely also explains an issue that Martin and I have both run into
where COUNT(*) queries occasionally resulted in incorrect counts.

Because the bug relies on the value of uninitialized memory, it's hard
to write a unit test for it, but I verified the change on the cluster
which previously crashed.

Change-Id: I1760847a52cb77adfbe56fd8054e0a7bbe9425ae
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/10037
Tested-by: jenkins
Reviewed-by: Alexander Behm <alex.behm@cloudera.com>