IMPALA-2711: Fix memory leak in Rand().
MathFunctions::RandPrepare() allocates a 4-bytes seed and
stores it in the FunctionContext's thread local state.
However, it was never freed. This change fixes the problem
by adding a close function for Rand() so it has a chance to
free the seed. A new test is also added to verify the fix.

Change-Id: Ibcc2e1ca0d052b86defe80aad471f9fdaac5a453
Reviewed-on: http://gerrit.cloudera.org:8080/1855
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 2c637b21bbcd76aa49560e0fce4411fdc5f86be9)
Reviewed-on: http://gerrit.cloudera.org:8080/2638
Reviewed-by: anujphadke <aphadke@cloudera.com>
Tested-by: Huaisi Xu <hxu@cloudera.com>