IMPALA-2859: don't use DCHECK in unit tests
DCHECKs are removed in release mode, so it is incorrect to use them in
unit tests that will be run in release mode. This caused promise-test to
randomly fail, since a function call was wrapped in a DCHECK and
therefore removed by the preprocessor in release mode.

This patch converts all DCHECKs in unit tests to either ASSERT (where
possible) or EXPECT (otherwise). In one place we use EXPECT_TRUE(... !=
...) because EXPECT_NE tries to take the address of a constant.

Change-Id: I1424b1a3154d748560d57d05cb8d69e6ee6e1923
Reviewed-on: http://gerrit.cloudera.org:8080/1890
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins