IMPALA-2592: DataStreamSender::Channel::CloseInternal() does not close the channel on an error.
CloseInternal() did not send an EOS RPC if there was an error in
sending the previous RPC or an error in flushing the current row
batch. It instead just logged the error. Due to this, the receiver
on the RPC channel remained open causing the query to hang in the
cluster.

This fix adds a new function called FlushFinal() to the class
DataSink. All the classes that inherit from DataSink
(DataStreamSender included) implement this function. It currently is
a no-op in the classes other than DataStreamSender; this will be
changed in following patches.

FlushFinal() can return a status and all the function calls that could
fail in DataStreamSender::Close() have been moved to FlushFinal()
leaving Close() as a pure teardown function which cannot fail.

Due to FlushFinal() being able to return an error status, the errors
are sent to the PlanFragmentExecutor which sends these errors in a
report via ReportStatusCb() to the coordinator. The coordinator then
cancels the query if it receives a report with an error in it. This
is how the fix helps in getting rid of the cluster hang.

Furthermore, a retry loop is added to send an RPC in
FragmentExecState::ReportStatusCb() in case the coordinator is
unreachable. If this RPC fails to be sent, there is a chance that the
cluster could hang. This will be addressed in a seperate patch.

Change-Id: Ia45d45dedb21c8b8abab266e140766eddfa363fb
Reviewed-on: http://gerrit.cloudera.org:8080/2205
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Henry Robinson <henry@cloudera.com>
(cherry picked from commit 9d36d6d98ff147b38736ec68c39987ad19bab1eb)
(cherry picked from commit 034a879070eccb541ce16331b982ecf1bad5c707)
(cherry picked from commit ecd9e21aa8026bbef78c945318e9818a671e7e2b)