IMPALA-2221: Fix the flaky test test_shell_commands.test_queries_closed.
The bug was that the test sometimes polled the wrong impalad for checking
metrics because ImpaladCluster() does not guarantee any particular order of
impalads, but the test was incorrectly relying on that.

The fix is to execute the shell commands on a fixed impalad, and use the
same impalad for polling metrics.

Change-Id: Iecb528be916d2900d7ec8a894bdef630250547da
Reviewed-on: http://gerrit.cloudera.org:8080/1974
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins