IMPALA-2837: make RuntimeState::GetUnreportedErrors() correctly reset reported errors
When remote fragments report runtime errors to the coordinator, they
call RuntimeState::GetUnreportedError(), which copies any unreported
errors and then clears them from the log so they are not reported
again. However, we weren't actually clearing the errors, so the same
errors would be reported over and over again. This rarely showed up in
our testing because our queries testing runtime error reporting are
very short, so they usually only send the errors a single time.

I'm working on creating a test case for this, but need to fix some
other bugs first.

Change-Id: I2d24e311c89e614c844643c7a91dfc8d8891cadb
Reviewed-on: http://gerrit.cloudera.org:8080/1996
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins