IMPALA-2093: Disallow NOT IN aggregate subqueries with a constant lhs expr
This commit fixes the issue where an incorrect rewrite is performed for the case of
uncorrelated aggregate subqueries participating in a NOT IN expr with a constant. With
this commit we detect cases like this and throw an AnalysisException.

Change-Id: Ia05ebc70ef310a0ec6f00ae6ae1c5de8fbd50240
Reviewed-on: http://gerrit.cloudera.org:8080/2007
Reviewed-by: Dimitris Tsirogiannis <dtsirogiannis@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit bf6fabd1029ffb3f087076a786ef9864509daef6)
Reviewed-on: http://gerrit.cloudera.org:8080/2158
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Bharath Vissapragada <bharathv@cloudera.com>