IMPALA-2940: fix leak of dictionaries in Parquet scanner
Per-split ColumnReader objects were stored in the RuntimeState
object pool and only freed at query end. The ColumnReader contains a
DictDecoder, which contains a vector of values for the last dictionary
page. This can add up to a significant amount of memory for very large
scans.

This commit adds a ObjectPool to HdfsScanner to be used for objects
that don't outlive a scanner. There are more objects associated with
scanners or ScannerContexts that should be freed earlier, but each
case needs to be validated individually, so this is left for a later
patch.

Also remove a dead function.

Change-Id: Icf8b155be17ec5f3ba17745f807bbd461202e3a0
Reviewed-on: http://gerrit.cloudera.org:8080/2029
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 6b434a5557fe03318a05062953e8ba08dd7918f6)
Reviewed-on: http://gerrit.cloudera.org:8080/2161
Reviewed-by: Juan Yu <jyu@cloudera.com>
Tested-by: Juan Yu <jyu@cloudera.com>