Misc MemPool cleanup
Delete dead code and move Clear() into .cc file (it is not called on
especially hot paths and the function is too large to benefit from
inlining). This is preparation for fixing IMPALA-2742 to reduce the
number of lines of code that must be considered.

Change-Id: I4a18b15e49c7c1618e32df7f956e4ba625aad144
Reviewed-on: http://gerrit.cloudera.org:8080/2070
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 4157b4bc095305796a8dd8f0cf7ae769828304a8)