IMPALA-2742: avoid unbounded MemPool growth with AcquireData()
This patch is a conservative fix for the possible unbounded growth in
MemPool size observed in IMPALA-2742. It preserves the doubling
behaviour up to a limit of 1MB, after which larger chunks are only
allocated if needed to fit the data. It also tracks the next chunk size
separately instead of relying on the size of the chunk at
current_chunk_idx_, which could be an arbitrary chunk in some cases
since the pool can rearrange free chunks if Allocate() is called after
Clear().

Remove logic for fixed chunk sizes that was only used in one test, and
update the test to account for chunk size doubling.

Also switch arguments to EXPECT_EQ() in tests so that actual versus
expected are in the right places.

Includes a unit test that reproduces the unbounded MemPool overhead
resulting from a series of small allocations intermixed with
AcquireData() calls.

Change-Id: I7f4e3d8663819b7b0bdc90bcaa74a40ec98bb4fd
Reviewed-on: http://gerrit.cloudera.org:8080/2096
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit d5eb2c9c0f634c3727730af6edc7f265d8f3395c)