IMPALA-2464: Disable per-row filters based on selectivity
This patch adds a simple heuristic to the Parquet scanner to disable a
filter if it's not being selective. If the fraction of rows rejected by
the filter drops below 10%, the filter is disabled. This is tested every
16K rows.

The ratio of rows rejected to total rows considered is controlled by
FLAGS_parquet_min_filter_reject_ratio.

Change-Id: Ic5d98c97061ccda7f0d4448f312b5a1618d81e56
Reviewed-on: http://gerrit.cloudera.org:8080/2116
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Henry Robinson <henry@cloudera.com>