IMPALA-2981: Fix coordinator waiting forever for remote fragments
The coordinator waits on all ExecRemoteFragment() calls to finish before
continuing to run the query. However, in one case, ExecRemoteFragment()
would not signal completion to the coordinator if the RPC had failed to
get a connection. This patch fixes that by adding a struct that will
always notify the barrier when the method exits.

Change-Id: Ic9176dc2e92996fe16b7f5ae64389b33113354c7
Reviewed-on: http://gerrit.cloudera.org:8080/2197
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins