IMPALA-2989: Admission Control crash on topic deletion
On bolt-80 we have seen a node crash when it receives a
topic deletion for which there is no corresponding entry in
the PoolStats::remote_stats_ map. The code could not handle
this case. I don't know exactly how this scenario occurred,
but we should avoid the crash for now and try to study this
further in cluster testing. It seemed to happen on a
particular node a few times in a row; the node would crash
and upon restart, it occurred again a number of times but
then stopped.

Change-Id: I1326579530ddfe0d8d02ddd46c3f01c6f8066845
Reviewed-on: http://gerrit.cloudera.org:8080/2177
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Internal Jenkins