Reduce size of cross-compiled IR
Disable the instnamer pass to reduce bitcode file size. Enable
inlining with a low threshold to inline many trivial functions in
the cross-compiled IR (e.g. scoped_ptr accessor functions).
Add additional IR_ALWAYS_INLINE annotations to trivial functions
that are called on hot paths.

This means that the functions will be inlined at compile time and
codegen has to do less work at runtime.

The IR module size is reduced from ~1.9MB to ~1.4MB with this
change and codegen time is slightly reduced. The number of functions is
reduced from 5698 to 3883.

Change-Id: Iea04ad2e4b365564ee71082657262485d3a85446
Reviewed-on: http://gerrit.cloudera.org:8080/2464
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins