IMPALA-3026: Mark filters with "local only" to avoid sending to the coordinator
With this commit runtime filters are marked as "local only" if they are produced by a
broadcast join and are applied by a scan node which is on the same plan fragment as the
join node. The BE can use that information to avoid sending these
filters to the coordinator, thereby reducing the number of filters that
the coordinator needs to process.

Change-Id: Id69c030b725f57758beecbd792bca28a761b912d
Reviewed-on: http://gerrit.cloudera.org:8080/2232
Reviewed-by: Dimitris Tsirogiannis <dtsirogiannis@cloudera.com>
Tested-by: Internal Jenkins