Validate the service while authorizing an internal connection
This patch validates the first component on an internal connection
seeking authorization. If the first component of the requesting user
does not match the first component of the principal validating it,
no authorization is given to that user.

This patch has been manually tested on a secure kerberized cluster.

Change-Id: I57f85ccda7b9004fce8206200eddfbc3c405cb7a
Reviewed-on: http://gerrit.cloudera.org:8080/2237
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 22f71f3e82151c1064bf2983dd3e9025c4cce5b3)
Reviewed-on: http://gerrit.cloudera.org:8080/2354
Reviewed-by: Juan Yu <jyu@cloudera.com>
Tested-by: Anuj Phadke <aphadke@cloudera.com>