IMPALA-3027: make DataStreamMgr::closed_stream_expirations_ a multimap
Under concurrency, it was possible for multiple streams to be added to
closed_stream_expirations_ with the same expiration time. This meant
that the clobbered stream would never be removed from
closed_stream_cache_. Fix this by using a multimap instead of a map.

I was able to reproduce this locally and verified that this patch
addresses the problem. The added DCHECK causes the local mini stress
test to fail without using multimap.

Change-Id: I551e3aa9c1c0a0eba9675d27af5d7bff1c6069d2
Reviewed-on: http://gerrit.cloudera.org:8080/2264
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: Internal Jenkins