IMPALA-3095: Add configurable whitelist of authorized internal principals
Although Impala now restricts the set of users authorized to access
internal APIs to the configured 'be' principal, there are a small number
of uses for allowing other named users to access the same APIs.

This patch adds a configurable whitelist of authorized users as a
command-line flag. By default, it is set to the HDFS user.

Change-Id: If828e86c99c3c398319953b1d3b33d5e3af200da
Reviewed-on: http://gerrit.cloudera.org:8080/2334
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Reviewed-by: Henry Robinson <henry@cloudera.com>
Reviewed-by: Silvius Rus <srus@cloudera.com>
Tested-by: Internal Jenkins