IMPALA-2982: lazy initialization of HBase connection
We shouldn't initialize a HBase connection if Impala will be run on a
cluster without Hbase. This started to bite us with a change to ZK or
HBase that resulted in periodic log messages when a bad HBase connection
was created. This patch changes the logic to only initialize the
question on demand, e.g. when a query is run against a HBase table.

Change-Id: I405f5cbfc62dd4a4aeb6f4019cac358376478884
Reviewed-on: http://gerrit.cloudera.org:8080/2345
Tested-by: Internal Jenkins
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>