IMPALA-3142: avoid spinlock in MemTracker::GcMemory()
When the process memory limit is exceeded we see excessive contention
for the gc_lock_ in the process memory tracker. A SpinLock is not a good
fit for this lock because we expect high contention (when the process
memory limit is exceeded, it is exceeded globally and many threads will
try to call into the function at the same time). boost::mutex will fall
back to a kernel-level mutex and block the process instead of spinning.

Change-Id: Ia9e91ea67ea924ba9b81e7376487838b365ca8ce
Reviewed-on: http://gerrit.cloudera.org:8080/2459
Reviewed-by: Michael Ho <kwho@cloudera.com>
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Reviewed-by: Silvius Rus <srus@cloudera.com>
Tested-by: Internal Jenkins