IMPALA-3243: Delete unneeded code that doesn't work on RHEL 5
The MaxCPUIndex() interface was added by Kudu, but the code that
initializes it doesn't work on RHEL 5. Impala doesn't depend on
MaxCpuIndex() at all, but the value is calculated and cached at startup
time so it still executes.  So, let's just delete this effectively
dead code that doesn't work on RHEL 5.

Change-Id: I82b71de60e5bf8abec18f981d13528991265908f
Reviewed-on: http://gerrit.cloudera.org:8080/2626
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins